Adds reducer-helpers for all protocols
