+++ packages/wallet/src/redux/protocols/reducer-helpers.ts
@@ -1,0 +1,38 @@
+import { SIGNATURE_SUCCESS, VALIDATION_SUCCESS } from 'magmo-wallet-client';
+import * as actions from '../actions';
+import { channelStateReducer } from '../channel-state/reducer';
+import { accumulateSideEffects } from '../outbox';
+import { SideEffects } from '../outbox/state';
+import { SharedData } from '../state';
+
+export const updateChannelState = (
+  sharedData: SharedData,
+  channelAction: actions.channel.ChannelAction,
+): SharedData => {
+  const newSharedData = { ...sharedData };
+  const updatedChannelState = channelStateReducer(newSharedData.channelState, channelAction);
+  newSharedData.channelState = updatedChannelState.state;
+  // TODO: Currently we need to filter out signature/validation messages that are meant to the app
+  // This might change based on whether protocol reducers or channel reducers craft commitments
+  const filteredSideEffects = filterOutSignatureMessages(updatedChannelState.sideEffects);
+  // App channel state may still generate side effects
+  newSharedData.outboxState = accumulateSideEffects(newSharedData.outboxState, filteredSideEffects);
+  return newSharedData;
+};
+
+export const filterOutSignatureMessages = (sideEffects?: SideEffects): SideEffects | undefined => {
+  if (sideEffects && sideEffects.messageOutbox) {
+    let messageArray = Array.isArray(sideEffects.messageOutbox)
+      ? sideEffects.messageOutbox
+      : [sideEffects.messageOutbox];
+    messageArray = messageArray.filter(
+      walletEvent =>
+        walletEvent.type !== VALIDATION_SUCCESS && walletEvent.type !== SIGNATURE_SUCCESS,
+    );
+    return {
+      ...sideEffects,
+      messageOutbox: messageArray,
+    };
+  }
+  return sideEffects;
+};
